Added lgt92 to general gps processor service

+++ project/backend-services/location-tracking-backend/domain/src/main/java/pt/sharespot/services/locationtrackingbackend/domain/sensor/gps/GPSDataDetails.java
@@ -2,7 +2,7 @@ package pt.sharespot.services.locationtrackingbackend.domain.sensor.gps;
 
 import pt.sharespot.services.locationtrackingbackend.domain.exceptions.NotValidException;
 
-public record GPSDataDetails(Long latitude, Long longitude) {
+public record GPSDataDetails(Double latitude, Double longitude) {
 
     public GPSDataDetails {
         if(latitude > 90 || latitude < -90) throw new NotValidException("Latitude must have a value between 90 and -90");
